Added Log writing

+++ model_selection.py
@@ -1,10 +1,11 @@
 import os
 import shutil
-from  Utility import getArgs
+from  Utility import getArgs, get_epoch
 
 def select_model():
     arg = getArgs()
-    threshold = 0.6
+    path = './numpy_bin'
+    threshold = 0.55
 
     eval_path = arg["eval_path"]
     model_path = arg["model_path"]
@@ -12,21 +13,32 @@ def select_model():
     n = len(os.listdir(eval_path))
     if n == 0:
         exit()
-    sum = .0
-
+    s = .0
+    w = 0
 
     for file in os.listdir(eval_path):
+        if os.path.isdir(f"{eval_path}/{file}"):
+            n-=1
+            continue
         with open(f"{eval_path}/{file}") as f:
-            sum += float(f.readline())
+            s += float(f.readline())
+            w += int(f.readline())
         os.remove(f"{eval_path}/{file}")
 
-    if sum >= threshold:
+    if s/n >= threshold:
+        print('Better Model found')
         current_model = model_path + '/current_model.ckpt'
         best_model = model_path + "/best_model.ckpt"
 
-        if os.path.isfile(model_path + '/best_model'):
-            os.remove(model_path + '/best_model')
+        if os.path.isfile(best_model):
+            os.remove(best_model)
         shutil.copy(current_model, best_model)
+    else:
+        print('No Model found')
+
+    with open('logs/main_log.out', "a") as f:
+        f.write(f"EPOCH: {get_epoch(path)-1}: Current Model won {w} games and has a WR of {s/n:.2f}\n\n")
+
